fix one-hot encoding error

+++ functions.py
@@ -6,7 +6,7 @@ def dice_loss(input, target):
     smooth = 1.
 
     iflat = input.view(-1)
-    tflat = target.contiguous().view(-1)
+    tflat = target.contiguous().view(-1).float()
     intersection = (iflat * tflat).sum()
 
     return 1 - ((2. * intersection + smooth) /
